Servidor/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-MySQLdb==0.2.0
greenlet==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector-python==8.0.24
mysqlclient==2.0.3
protobuf==3.15.8
PyJWT==1.4.2
PyMySQL==1.0.2
python-dotenv==0.17.1
six==1.16.0
SQLAlchemy==1.4.13
typing==3.7.4.3
uWSGI==2.0.19.1
Werkzeug==1.0.1
pycrypto==2.6.1
pycryptodome==3.9.7
